requirements.txt
Django
django-rest-framework
djangorestframework
Pillow
pip-check
pip-upgrader
whitenoise
mysqlclient
reportlab
pipreqs
python-decouple
django-cors-headers
djangorestframework-simplejwt
xhtml2pdf